+++ lib/govuk_publishing_components/presenters/shared_helper.rb
@@ -1,13 +1,20 @@
 module GovukPublishingComponents
   module Presenters
     class SharedHelper
-      attr_reader :options, :margin_top, :margin_bottom, :heading_level
+      attr_reader :options, :margin_top, :margin_bottom, :heading_level, :classes
 
       def initialize(local_assigns)
         @options = local_assigns
         @margin_top = @options[:margin_top] || nil
         @margin_bottom = @options[:margin_bottom] || 3
         @heading_level = @options[:heading_level] || 2
+
+        if local_assigns.include?(:classes)
+          @classes = local_assigns[:classes].split(" ")
+          unless @classes.all? { |c| c.start_with?("js-") }
+            raise(ArgumentError, "Passed classes must be prefixed with `js-`")
+          end
+        end
       end
 
       def get_margin_top